+++ olympia/gm.c
@@ -58,7 +58,7 @@ gm_show_skill_use_counts(int pl)
 	}
 	next_skill;
 
-	qsort(l, ilist_len(l), sizeof(int), skill_use_comp);
+	qsort(l, (size_t)ilist_len(l), sizeof(int), skill_use_comp);
 
 	out_path = MASTER;
 	out_alt_who = OUT_LORE;
@@ -131,7 +131,7 @@ gm_show_skills_known(int pl)
 	}
 	next_skill;
 
-	qsort(l, ilist_len(l), sizeof(int), skills_known_comp);
+	qsort(l, (size_t)ilist_len(l), sizeof(int), skills_known_comp);
 
 	out_path = MASTER;
 	out_alt_who = OUT_LORE;
@@ -694,7 +694,7 @@ gm_land_stats(int pl)
 	}
 	next_loc;
 
-	qsort(l, ilist_len(l), sizeof(int), region_occupy_comp);
+	qsort(l, (size_t)ilist_len(l), sizeof(int), region_occupy_comp);
 
 	out_path = MASTER;
 	out_alt_who = OUT_LORE;
@@ -748,7 +748,7 @@ gm_faction_wealth(int pl)
 	}
 	next_player;
 
-	qsort(l, ilist_len(l), sizeof(int), wealth_list_comp);
+	qsort(l, (size_t)ilist_len(l), sizeof(int), wealth_list_comp);
 
 	out_path = MASTER;
 	out_alt_who = OUT_LORE;
@@ -801,7 +801,7 @@ gm_nobles_list(int pl)
 	}
 	next_player;
 
-	qsort(l, ilist_len(l), sizeof(int), nobles_list_comp);
+	qsort(l, (size_t)ilist_len(l), sizeof(int), nobles_list_comp);
 
 	out_path = MASTER;
 	out_alt_who = OUT_LORE;
